+++ async_upnp_client/advertisement.py
@@ -3,7 +3,6 @@
 
 import asyncio
 import logging
-import sys
 from asyncio.events import AbstractEventLoop
 from asyncio.transports import BaseTransport, DatagramTransport
 from typing import Awaitable, Callable, Optional, Union
@@ -27,9 +26,15 @@ class SsdpAdvertisementListener:
 
     def __init__(
         self,
-        on_alive: Optional[Callable[[SsdpHeaders], Awaitable]] = None,
-        on_byebye: Optional[Callable[[SsdpHeaders], Awaitable]] = None,
-        on_update: Optional[Callable[[SsdpHeaders], Awaitable]] = None,
+        on_alive: Optional[
+            Callable[[SsdpHeaders, AddressTupleVXType], Awaitable]
+        ] = None,
+        on_byebye: Optional[
+            Callable[[SsdpHeaders, AddressTupleVXType], Awaitable]
+        ] = None,
+        on_update: Optional[
+            Callable[[SsdpHeaders, AddressTupleVXType], Awaitable]
+        ] = None,
         source: Union[AddressTupleVXType, IPvXAddress, None] = None,
         target: Union[AddressTupleVXType, IPvXAddress, None] = None,
         loop: Optional[AbstractEventLoop] = None,
@@ -44,13 +49,17 @@ class SsdpAdvertisementListener:
         self._loop: AbstractEventLoop = loop or asyncio.get_event_loop()
         self._transport: Optional[BaseTransport] = None
 
-    async def _async_on_data(self, request_line: str, headers: SsdpHeaders) -> None:
+    async def _async_on_data(
+        self, request_line: str, headers: SsdpHeaders, addr: AddressTupleVXType
+    ) -> None:
         """Handle data."""
         if headers.get("MAN") == SSDP_DISCOVER:
             # Ignore discover packets.
             return
         if "NTS" not in headers:
-            _LOGGER.debug("Got non-advertisement packet: %s, %s", request_line, headers)
+            _LOGGER.debug(
+                "Ignoring non-advertisement packet: %s, %s", request_line, headers
+            )
             return
 
         _LOGGER.debug("Received advertisement, USN: %s", headers.get("USN", "<no USN>"))
@@ -58,15 +67,15 @@ class SsdpAdvertisementListener:
         headers["_source"] = SsdpSource.ADVERTISEMENT
         notification_sub_type = headers["NTS"]
         if notification_sub_type == NotificationSubType.SSDP_ALIVE and self.on_alive:
-            await self.on_alive(headers)
+            await self.on_alive(headers, addr)
         elif (
             notification_sub_type == NotificationSubType.SSDP_BYEBYE and self.on_byebye
         ):
-            await self.on_byebye(headers)
+            await self.on_byebye(headers, addr)
         elif (
             notification_sub_type == NotificationSubType.SSDP_UPDATE and self.on_update
         ):
-            await self.on_update(headers)
+            await self.on_update(headers, addr)
 
     async def _async_on_connect(self, transport: DatagramTransport) -> None:
         self._transport = transport
@@ -76,10 +85,9 @@ class SsdpAdvertisementListener:
         _LOGGER.debug("Start listening for advertisements")
 
         # Construct a socket for use with this pairs of endpoints.
-        sock, sock_source, sock_target = get_ssdp_socket(self.source, self.target)
-        address = sock_source if sys.platform == "win32" else sock_target
-        _LOGGER.debug("Binding to address: %s", address)
-        sock.bind(address)
+        sock, _, sock_target = get_ssdp_socket(self.source, self.target)
+        _LOGGER.debug("Binding to address: %s", sock_target)
+        sock.bind(sock_target)
 
         # Create protocol and send discovery packet.
         await self._loop.create_datagram_endpoint(