+++ tests/test_advertisement.py
@@ -12,6 +12,7 @@ from async_upnp_client.advertisement import SsdpAdvertisementListener
 from async_upnp_client.utils import CaseInsensitiveDict
 
 from .common import (
+    ADDR,
     ADVERTISEMENT_HEADERS_DEFAULT,
     ADVERTISEMENT_REQUEST_LINE,
     SEARCH_HEADERS_DEFAULT,
@@ -31,9 +32,9 @@ async def test_receive_ssdp_alive() -> None:
     )
     headers = CaseInsensitiveDict(ADVERTISEMENT_HEADERS_DEFAULT)
     headers["NTS"] = "ssdp:alive"
-    await listener._async_on_data(ADVERTISEMENT_REQUEST_LINE, headers)
+    await listener._async_on_data(ADVERTISEMENT_REQUEST_LINE, headers, ADDR)
 
-    on_alive.assert_called_with(headers)
+    on_alive.assert_called_with(headers, ADDR)
     on_byebye.assert_not_called()
     on_update.assert_not_called()
 
@@ -50,10 +51,10 @@ async def test_receive_ssdp_byebye() -> None:
     )
     headers = CaseInsensitiveDict(ADVERTISEMENT_HEADERS_DEFAULT)
     headers["NTS"] = "ssdp:byebye"
-    await listener._async_on_data(ADVERTISEMENT_REQUEST_LINE, headers)
+    await listener._async_on_data(ADVERTISEMENT_REQUEST_LINE, headers, ADDR)
 
     on_alive.assert_not_called()
-    on_byebye.assert_called_with(headers)
+    on_byebye.assert_called_with(headers, ADDR)
     on_update.assert_not_called()
 
 
@@ -69,11 +70,11 @@ async def test_receive_ssdp_update() -> None:
     )
     headers = CaseInsensitiveDict(ADVERTISEMENT_HEADERS_DEFAULT)
     headers["NTS"] = "ssdp:update"
-    await listener._async_on_data(ADVERTISEMENT_REQUEST_LINE, headers)
+    await listener._async_on_data(ADVERTISEMENT_REQUEST_LINE, headers, ADDR)
 
     on_alive.assert_not_called()
     on_byebye.assert_not_called()
-    on_update.assert_called_with(headers)
+    on_update.assert_called_with(headers, ADDR)
 
 
 @pytest.mark.asyncio
@@ -87,7 +88,7 @@ async def test_receive_ssdp_search_response() -> None:
         on_alive=on_alive, on_byebye=on_byebye, on_update=on_update
     )
     headers = CaseInsensitiveDict(SEARCH_HEADERS_DEFAULT)
-    await listener._async_on_data(SEARCH_REQUEST_LINE, headers)
+    await listener._async_on_data(SEARCH_REQUEST_LINE, headers, ADDR)
 
     on_alive.assert_not_called()
     on_byebye.assert_not_called()